Nest education details in `properties`, update tests
