+++ src/schemas/FEEDBACK-TOOL/schema.js
@@ -216,83 +216,85 @@ let schema = {
         'school',
         'programs'
       ],
-      school: {
-        type: 'object',
-        required: ['address', 'name'],
-        properties: {
-          address: {
-            type: 'object',
-            oneOf: schoolAddress
-          },
-          name: { // Type: text (255)
-            type: 'string',
-            minLength: 1,
-            maxLength: 255
+      properties: {
+        school: {
+          type: 'object',
+          required: ['address', 'name'],
+          properties: {
+            address: {
+              type: 'object',
+              oneOf: schoolAddress
+            },
+            name: { // Type: text (255)
+              type: 'string',
+              minLength: 1,
+              maxLength: 255
+            }
           }
-        }
-      },
-      programs: { // TRANSLATE into array of strings
-        type: 'object', // FE validation requires at least one selected
-        properties: {
-          'Post-9/11 Ch 33': {
-            type: 'boolean',
-            default: false,
-            title: 'Post-9/11 GI Bill (Chapter 33)'
-          },
-          'MGIB-AD Ch 30': {
-            type: 'boolean',
-            default: false,
-            title: 'Montgomery GI Bill - Active Duty (MGIB-AD, Chapter 30)'
-          },
-          'MGIB-SR Ch 1606': {
-            type: 'boolean',
-            default: false,
-            title: 'Montgomery GI Bill - Selected Reserve (MGIB-SR, Chapter 1606)'
-          },
-          TATU: {
-            type: 'boolean',
-            default: false,
-            title: 'Tuition Assistance Top-Up'
-          },
-          REAP: {
-            type: 'boolean',
-            default: false,
-            title: 'Reserve Educational Assistance Program (REAP) (Chapter 1607)'
-          },
-          'DEA Ch 35': {
-            type: 'boolean',
-            default: false,
-            title: 'Survivors’ and Dependents’ Assistance (DEA) (Chapter 35)'
-          },
-          'VRE Ch 31': {
-            type: 'boolean',
-            default: false,
-            title: 'Vocational Rehabilitation and Employment (VR&E) (Chapter 31)'
+        },
+        programs: { // TRANSLATE into array of strings
+          type: 'object', // FE validation requires at least one selected
+          properties: {
+            'Post-9/11 Ch 33': {
+              type: 'boolean',
+              default: false,
+              title: 'Post-9/11 GI Bill (Chapter 33)'
+            },
+            'MGIB-AD Ch 30': {
+              type: 'boolean',
+              default: false,
+              title: 'Montgomery GI Bill - Active Duty (MGIB-AD, Chapter 30)'
+            },
+            'MGIB-SR Ch 1606': {
+              type: 'boolean',
+              default: false,
+              title: 'Montgomery GI Bill - Selected Reserve (MGIB-SR, Chapter 1606)'
+            },
+            TATU: {
+              type: 'boolean',
+              default: false,
+              title: 'Tuition Assistance Top-Up'
+            },
+            REAP: {
+              type: 'boolean',
+              default: false,
+              title: 'Reserve Educational Assistance Program (REAP) (Chapter 1607)'
+            },
+            'DEA Ch 35': {
+              type: 'boolean',
+              default: false,
+              title: 'Survivors’ and Dependents’ Assistance (DEA) (Chapter 35)'
+            },
+            'VRE Ch 31': {
+              type: 'boolean',
+              default: false,
+              title: 'Vocational Rehabilitation and Employment (VR&E) (Chapter 31)'
+            }
           }
-        }
-      },
-      assistance: { // TRANSLATE into array of strings
-        type: 'object',
-        properties: {
-          TA: {
-            type: 'boolean',
-            default: false,
-            title: 'Federal Tuition Assistance (TA)'
-          },
-          'TA-AGR': {
-            type: 'boolean',
-            default: false,
-            title: 'State-funded Tuition Assistance (TA) for Servicemembers on Active Guard and Reserve (AGR) duties'
-          },
-          MyCAA: {
-            type: 'boolean',
-            default: false,
-            title: 'Military Spouse Career Advancement Accounts (MyCAA)'
-          },
-          FFA: {
-            type: 'boolean',
-            default: false,
-            title: 'Federal financial aid'
+        },
+        assistance: { // TRANSLATE into array of strings
+          type: 'object',
+          properties: {
+            TA: {
+              type: 'boolean',
+              default: false,
+              title: 'Federal Tuition Assistance (TA)'
+            },
+            'TA-AGR': {
+              type: 'boolean',
+              default: false,
+              title: 'State-funded Tuition Assistance (TA) for Servicemembers on Active Guard and Reserve (AGR) duties'
+            },
+            MyCAA: {
+              type: 'boolean',
+              default: false,
+              title: 'Military Spouse Career Advancement Accounts (MyCAA)'
+            },
+            FFA: {
+              type: 'boolean',
+              default: false,
+              title: 'Federal financial aid'
+            }
           }
         }
       }
